Most of the updates Danny wanted
