+++ shoebox/app/assets/javascripts/website/onboarding.js
@@ -60,10 +60,6 @@ $(function() {
     });
     
   });
-  
-  
-
-
 
 });//end jQuery
