+++ shoebox/app/views/website/install.js
@@ -1,0 +1,7 @@
+$(function() {
+  $("#chrome-install").click(function() {
+    chrome.webstore.install(function() {
+      alert("Installed!")
+    })  
+  });
+});
