+++ shoebox/app/views/website/inviteFriends.scala.html
@@ -56,7 +56,7 @@
             </li>
           } else {
             <li class="clearfix">
-              <form action="@com.keepit.controllers.website.routes.HomeController.inviteConnection" method="POST">
+              <form action="@com.keepit.controllers.website.routes.InviteController.inviteConnection" method="POST">
                 <img src="@routes.Assets.at("images/default_avatar.png")" width="75" height="75" class="avatar">
                 <div class="avatar_info">
                   <input type="text" class="invite_field" value="Select a facebook friend to invite">
