--- shoebox/app/views/website/onboarding/extInstall.scala.html
@@ -1,30 +1,0 @@
-@(user: User, email: String)
-
-@views.html.website.helper.main(title = "Installing Kifi", hiddenElements = Seq("inviteFriends")) {
-  <div class="user_avatar">
-  <img src="@com.keepit.controllers.assets.routes.UserPictureController.get(112, user.externalId)" width="112" height="112">
-  </div>
-
-  <section class="content clearfix">
-    <header class="content_header">
-        <h2>Yay Effi, you’re in!</h2>
-        <h3>Welcome to the KiFi Private Beta</h3>
-    </header>
-
-    <h2 class="blue text_center">Start using kifi by installing the KiFi extention</h2>
-
-    <div class="install_extensions clearfix">
-      <div class="container left">
-          <p>For Google Chrome</p>
-            <a class="btn blue" href="#">Install Extension</a>
-      </div>
-      <div class="container right">
-          <p>For Google Chrome</p>
-            <a class="btn blue" href="#">Install Extension</a>
-      </div>                    
-    </div>                
-
-  </section>
-
-
-}
