+++ shoebox/app/views/website/userHome.scala.html
@@ -27,11 +27,10 @@
     <h2>What would you like to do today?</h2>
 
     <ul class="links_list">
-      <li><a href="#">Invite friends to the Private beta</a></li>
-        <li><a href="#">Watch the kifi demo video</a></li>
-        <li><a href="#">Send Feedback</a></li>
-        <li><a href="#">I want to know more about you guys</a></li>
-        <li><a href="#">Nothing, just like watching this page for no reason</a></li>
+      <!--<li><a href="/invite">Invite friends to the Private beta</a></li>-->
+      <li><a href="/install">Install the KiFi extension on this computer</a></li>
+      <li><a href="mailto:feedback@@kifi.com">Send Feedback</a></li>
+      <li><a href="http://www.42go.com/team.html">I want to know more about you guys</a></li>
     </ul>
   </section>
 }
